requirements.txt
streamlit
pandas
requests
biopython
GEOparse
pysradb
stringdb
# PRIDE, MetaboLights APIs halna baaki later ig